requirements.txt
luma.oled
adafruit-pca9685
adafruit-circuitpython-motor
rpi_ws281x
mpu6050-raspberrypi
websockets
numpy
opencv-contrib-python >= 3.4.3.18
imutils
zmq
pybase64
psutil
smbus
adafruit-blinka
adafruit-circuitpython-pca9685
picamera2
opencv-python
adafruit-circuitpython-board
fastapi
starlette
uvicorn
python-dotenv